requirements.txt
pandas
pytest
streamlit
spotipy
anthropic
python-dotenv
